+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## [Unreleased]
+### Added
 * Add new CA Death option to the unprocessed record filter (#75)
 
 ## 5.5.0 / 2021-03-16